% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ams_iig_equations.R
\name{calculate_thermal_energy_iig}
\alias{calculate_thermal_energy_iig}
\title{Internal helper to convert non-renewable biomass into thermal energy}
\usage{
calculate_thermal_energy_iig(
  non_renewable_biomass,
  biomass_col,
  energy_data,
  ncv_col,
  group_cols,
  output_col
)
}
\description{
Internal helper to convert non-renewable biomass into thermal energy
}
\keyword{internal}
